added reload after delete, created model confirmation

+++ app/scripts/controllers/entrance.js
@@ -44,8 +44,8 @@ angular.module('viLoggedClientApp')
         }
       })
   })
-  .controller('CompanyEntranceCtrl', function($scope, entranceService, $rootScope) {
-    $rootScope.busy = false;
+  .controller('CompanyEntranceCtrl', function($scope, entranceService, $rootScope, notificationService) {
+
     $scope.entrance = [];
 
     $scope.pagination = {
@@ -54,29 +54,43 @@ angular.module('viLoggedClientApp')
       currentPage: 1
     };
 
+    getEntrance();
+
     $scope.deleteEntrance = function(id) {
+      var dialogParams = {
+        modalHeader: 'Delete Entrance',
+        modalBodyText: 'are you sure you want to delete the following?'
+      };
+
+      notificationService.modal.confirm(dialogParams)
+        .then(function() {
+          $rootScope.busy = true;
+          entranceService.remove(id)
+            .then(function(response) {
+              $rootScope.busy = false;
+              getEntrance();
+            })
+            .catch(function(reason) {
+              $rootScope.busy = false;
+              notificationService.setTimeOutNotification(reason);
+            });
+        });
+    };
+    function getEntrance() {
       $rootScope.busy = true;
-      entranceService.remove(id)
+      entranceService.all()
         .then(function(response) {
           $rootScope.busy = false;
+          $scope.entrance = response;
+          $scope.pagination.totalItems = $scope.entrance.length;
+          $scope.pagination.numberOfPages = Math.ceil($scope.pagination.totalItems / $scope.pagination.itemsPerPage);
         })
         .catch(function(reason) {
           $rootScope.busy = false;
 
-        })
-    };
+        });
+    }
 
-    entranceService.all()
-      .then(function(response) {
-        $rootScope.busy = false;
-        $scope.entrance = response;
-        $scope.pagination.totalItems = $scope.entrance.length;
-        $scope.pagination.numberOfPages = Math.ceil($scope.pagination.totalItems / $scope.pagination.itemsPerPage);
-      })
-      .catch(function(reason) {
-        $rootScope.busy = false;
-
-      });
   })
   .controller('EntranceFormCtrl', function($scope, $state, $stateParams, entranceService, $rootScope, notificationService) {
     $scope.entrance = {};
